Finished deleteAttribute

+++ UserInterface.java
@@ -1,4 +1,5 @@
 import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.Scanner;
 
 public class UserInterface {
@@ -226,23 +227,28 @@ public class UserInterface {
         for (int i = 0; i < createdClasses.size(); i++) {
             if (createdClasses.get(i).getName().toLowerCase().equals(className)) {
                 // check if attributes is empty, if not do below
-                System.out.println("Which attribute would you like to delete?");
-                // list attributes
-                for(int j = 0; j < createdClasses.get(i).getAttributes().size(); i++){
-                    int counter = j + 1;
-                    System.out.println(counter + ".) " + createdClasses.get(i).getAttributes().get(j));
-                }
-                String attribute = kb.nextLine();
-                // find attribute using loop
-                System.out.println("Are you sure you want to delete " + attribute + "? Please enter yes or no.");
-                String answer = kb.nextLine().toLowerCase();
-                //stick the part below in a loop so if an incorrect input is entered, it'll re-prompt
-                if (answer.equals("yes")) {
-                    createdClasses.get(i).deleteAttribute(attribute);
-                } else if (answer.equals("no")) {
-                    System.out.println("Canceled");
+                LinkedList tempAttList = createdClasses.get(i).getAttributes();
+                if(!(tempAttList==null)) {
+                    System.out.println("Which attribute would you like to delete?");
+                    // lists attributes
+                    for (int j = 0; j < createdClasses.get(i).getAttributes().size(); i++) {
+                        int counter = j + 1;
+                        System.out.println(counter + ".) " + createdClasses.get(i).getAttributes().get(j));
+                    }
+                    String attribute = kb.nextLine();
+                    // find attribute using loop
+                    System.out.println("Are you sure you want to delete " + attribute + "? Please enter yes or no.");
+                    String answer = kb.nextLine().toLowerCase();
+                    //stick the part below in a loop so if an incorrect input is entered, it'll re-prompt
+                    if (answer.equals("yes")) {
+                        createdClasses.get(i).deleteAttribute(attribute);
+                    } else if (answer.equals("no")) {
+                        System.out.println("Canceled");
+                    } else {
+                        System.out.println("That is not a valid input");
+                    }
                 } else {
-                    System.out.println("That is not a valid input");
+                    System.out.println("No attributes found for class " + className);
                 }
             }
         }
